🏷️ move swagger-client types to d.ts
