+++ app/javascript/src/Types/SwaggerTypes.ts
@@ -1,5 +1,4 @@
 /* eslint-disable @typescript-eslint/naming-convention */
-import type { Request, Response, SupportedHTTPMethods } from 'swagger-ui'
 
 export type AccountData = Record<string, {
   name: string;
@@ -17,26 +16,6 @@ export interface ApiDocsService {
 export interface ApiDocsServices {
   host: string;
   apis: ApiDocsService[];
-}
-
-export interface ExecuteData {
-  contextUrl: string;
-  fetch: (arg: unknown) => unknown;
-  method: SupportedHTTPMethods;
-  operation: unknown;
-  operationId: string;
-  parameters: unknown;
-  pathName: string;
-  requestBody?: unknown;
-  requestContentType: string;
-  requestInterceptor?: ((request: Request) => Promise<Request> | Request) | undefined;
-  responseContentType: string;
-  responseInterceptor?: ((response: Response) => Promise<Response> | Response) | undefined;
-  scheme: string;
-  securities: unknown;
-  server: string;
-  serverVariables: unknown;
-  spec: unknown;
 }
 
 export interface BackendApiTransaction extends BodyValueObject {
